Decode THUMB instructions, add partial support for THUMB instructions
Added some offsets as determined by GHIDRA when branching or calculating
based on the Program Counter. This is required as I'm not incrementing
the program counter until the end of the instruction, to avoid having to
deal with instruction pipelining.

This is a bodge and should be removed once I understand better how to
implement the instruction pipeline.

Also added a function to allow pausing operation for user input.